fix: make contact-info.json optional
Since contact-info.json is not committed, it must be treated as
optional. If the file is not present, use an empty object `{}` instead.

Without _any_ contact info, however, I will be unreachable. Include some
public contact details in `pug.config.js` so that the resume published
to gh-pages includes some means of contacting me.