fix: reload dev server whenever the pug locals files change
Whenever pug.config.js or pug.config.yaml change, we want the dev server
to restart.

This is because parcel will not hot-reload on changes to the
`pug.config.js` file. We must restart the server then reload the page in
order to see content changes. `nodemon` will restart the server for us,
so now we only need to reload the page, which isn't a huge drag.